#141 Moved gold to CombatStatsComponent.java
