+++ source/core/src/main/com/csse3200/game/components/CombatStatsComponent.java
@@ -20,6 +20,12 @@ public class CombatStatsComponent extends Component {
     setBaseAttack(baseAttack);
   }
 
+  public CombatStatsComponent(int health, int baseAttack, int gold) {
+    setHealth(health);
+    setBaseAttack(baseAttack);
+    setGold(gold);
+  }
+
   /**
    * Returns true if the entity's has 0 health, otherwise false.
    *
@@ -88,5 +94,35 @@ public class CombatStatsComponent extends Component {
   public void hit(CombatStatsComponent attacker) {
     int newHealth = getHealth() - attacker.getBaseAttack();
     setHealth(newHealth);
+  }
+
+  /**
+   * Sets the entity's gold. Gold has a minimum bound of 0.
+   *
+   * @param gold gold
+   */
+  public void setGold(int gold) {
+    this.gold = Math.max(gold, 0);
+    if (entity != null) {
+      entity.getEvents().trigger("updateGold", this.gold);
+    }
+  }
+
+  /**
+   * Adds to the player's gold. The amount added can be negative.
+   *
+   * @param gold gold to add
+   */
+  public void addGold(int gold) {
+    setGold(this.gold + gold);
+  }
+
+  /**
+   * Returns the entity's gold.
+   *
+   * @return entity's gold
+   */
+  public int getGold() {
+    return gold;
   }
 }
