+++ source/core/src/main/com/csse3200/game/components/player/InventoryComponent.java
@@ -21,14 +21,12 @@ public class InventoryComponent extends Component {
   private final String sizeException = "Invalid size parameter. Must be an integer > 0.";
   private final String itemException = "Index in Inventory already occupied by an Item.";
   private final String nullException = "Index in Inventory does not contain an Item.";
-  private int gold;
 
   /**
    * Creates inventory component
    * @param capacity the players inventory size
-   * @param gold the players gold
    */
-  public InventoryComponent(int capacity, int gold) {
+  public InventoryComponent(int capacity) {
     setCapacity(capacity);
     items = new ArrayList<>(capacity);
     for (int i = 0; i < capacity; i++) {        
@@ -36,7 +34,6 @@ public class InventoryComponent extends Component {
     }
     size = 0;
     setSelected(0);
-    setGold(gold);
   }
 
   /**
@@ -69,36 +66,6 @@ public class InventoryComponent extends Component {
         throw new IllegalArgumentException(sizeException);
       }
       this.capacity = newCapacity;
-  }
-
-  /**
-   * Sets the entity's gold. Gold has a minimum bound of 0.
-   *
-   * @param gold gold
-   */
-  public void setGold(int gold) {
-    this.gold = Math.max(gold, 0);
-    if (entity != null) {
-      entity.getEvents().trigger("updateGold", this.gold);
-    }
-  }
-
-  /**
-   * Adds to the player's gold. The amount added can be negative.
-   *
-   * @param gold gold to add
-   */
-  public void addGold(int gold) {
-    setGold(this.gold + gold);
-  }
-
-  /**
-   * Returns the entity's gold.
-   *
-   * @return entity's gold
-   */
-  public int getGold() {
-    return gold;
   }
 
   /**
