+++ source/core/src/main/com/csse3200/game/entities/factories/PlayerFactory.java
@@ -46,8 +46,8 @@ public class PlayerFactory {
             .addComponent(new ColliderComponent())
             .addComponent(new HitboxComponent().setLayer(PhysicsLayer.PLAYER))
             .addComponent(new PlayerActions())
-            .addComponent(new CombatStatsComponent(stats.health, stats.baseAttack))
-            .addComponent(new InventoryComponent(stats.inventorySize, stats.gold))
+            .addComponent(new CombatStatsComponent(stats.health, stats.baseAttack, stats.gold))
+            .addComponent(new InventoryComponent(stats.inventorySize))
             .addComponent(new InventoryDisplay())
             .addComponent(inputComponent)
             .addComponent(new TooltipsDisplay())
